requirements.txt
python-telegram-bot==20.6
gspread==5.12.0
google-auth==2.34.0
google-auth-oauthlib==1.2.2
httpx==0.25.2
requests==2.32.3
python-dotenv==1.0.1
openai==1.17.0

